Support universal equality of unsigned numeric types (#3584)
* Make scala.scananative.unsigned types and Size extend ScalaNumber
* Adapt scala.runtime.BoxesRunTime to handle `==` with unsigned types
* Port logic for handling unversal equality from JVM backend to the compiler plugin.
Use BoxesRunTime.equals methods instead of `_Object.__scala_==`
* Remove no longer used `__scala_==` methods
* Add tests for universal equality of unsigned types
* Apply universal equality with unsigned types whenever possible
* Add docs entry for universal equality for unsiged types
* Fix compilation after rebase
* Fix compilation of Scala 2.12
* Update minimal required symbols metrics and fix typos
* Disable one failing case in UnsisgnedUniversalEqualityTest for Size on 32bit arch
- Fix CI for usize on linux-x86
- Fix compilation warning

+++ javalib/src/main/scala/java/lang/process/WindowsProcess.scala
@@ -323,9 +323,9 @@ object WindowsProcess {
     list.add("")
     val totalSize = list.scalaOps.foldLeft(0)(_ + _.size + 1) - 1
     val blockEnd = block + totalSize
-    assert(!blockEnd == 0.toUShort, s"not null terminated got ${!blockEnd}")
+    assert(!blockEnd == 0, s"not null terminated got ${!blockEnd}")
     assert(
-      !(blockEnd - 1) == 0.toUShort,
+      !(blockEnd - 1) == 0,
       s"not null terminated -1, got ${!(blockEnd - 1)}"
     )
